For #38645 Add drop frame support
Adds drop frame support to the `timecode` and `edl` modules for 29.97 and 59.94 frame rates. Drop frame can be designated by using a drop_frame parameter and/or by using drop frame notation in the timecode string (using ; , or . as the delimited between seconds and frames). Calculations to and from frames take drop frame into account and adjust the results as needed. Several good reference websites are in the docs as well as additional information.

Additional changes:

* Updated license.
* Added docstring updates and docs folder for sphinx.
* Fixed wrong record_out returned for transition EditEvents.
* Fixed clip name for EditEvents with transitions.
* Added/updated tests and reorganized test resources folder.